refactor(organon): extract energeia::steward submodule (#2642)
Move epitropos tool definition and EpitroposExecutor to steward.rs.
