+++ crates/organon/src/builtins/energeia/mod.rs
@@ -14,6 +14,7 @@
 mod dispatch;
 mod qa;
 mod shared;
+mod steward;
 
 pub use shared::EnergeiaServices;
 
@@ -40,100 +41,6 @@ use crate::types::{
 };
 
 use shared::{opt_bool, opt_str, opt_u64, require_str, to_json_text};
-
-// ── epitropos (ἐπίτροπος — steward) ───────────────────────────────────────
-
-fn epitropos_def() -> ToolDef {
-    ToolDef {
-        name: ToolName::from_static("epitropos"),
-        description: "CI steward: monitor pull requests, auto-merge passing PRs, \
-            queue failing PRs for repair. Runs as a polling loop unless `once` is set."
-            .to_owned(),
-        extended_description: None,
-        input_schema: InputSchema {
-            properties: IndexMap::from([
-                (
-                    "project".to_owned(),
-                    PropertyDef {
-                        property_type: PropertyType::String,
-                        description: "GitHub project slug (owner/repo)".to_owned(),
-                        enum_values: None,
-                        default: None,
-                    },
-                ),
-                (
-                    "once".to_owned(),
-                    PropertyDef {
-                        property_type: PropertyType::Boolean,
-                        description: "Run a single classification pass instead of a polling loop \
-                            (default: false)"
-                            .to_owned(),
-                        enum_values: None,
-                        default: Some(serde_json::json!(false)),
-                    },
-                ),
-                (
-                    "dry_run".to_owned(),
-                    PropertyDef {
-                        property_type: PropertyType::Boolean,
-                        description: "Classify PRs without merging or queuing repairs \
-                            (default: false)"
-                            .to_owned(),
-                        enum_values: None,
-                        default: Some(serde_json::json!(false)),
-                    },
-                ),
-            ]),
-            required: vec!["project".to_owned()],
-        },
-        category: ToolCategory::Agent,
-        reversibility: Reversibility::Irreversible,
-        auto_activate: false,
-    }
-}
-
-struct EpitroposExecutor;
-
-impl ToolExecutor for EpitroposExecutor {
-    fn execute<'a>(
-        &'a self,
-        input: &'a ToolInput,
-        _ctx: &'a ToolContext,
-    ) -> Pin<Box<dyn Future<Output = Result<ToolResult>> + Send + 'a>> {
-        Box::pin(async move {
-            let args = &input.arguments;
-
-            let project = match require_str(args, "project") {
-                Ok(s) => s,
-                Err(e) => return Ok(e),
-            };
-            let once = opt_bool(args, "once").unwrap_or(false);
-            let dry_run = opt_bool(args, "dry_run").unwrap_or(false);
-
-            let mut config = StewardConfig::new(project.to_owned());
-            config.once = once;
-            config.dry_run = dry_run;
-
-            // WHY: Always use run_once in tool context — a polling loop would block
-            // the tool executor indefinitely. Callers that need the polling loop
-            // should schedule a recurring trigger instead.
-            let result = run_once(&config).await;
-
-            let output = serde_json::json!({
-                "project": project,
-                "dry_run": dry_run,
-                "classified_count": result.classified.len(),
-                "merged_count": result.merged.len(),
-                "needs_fix_count": result.needs_fix.len(),
-                "blocked_count": result.blocked.len(),
-                "main_ci_status": format!("{:?}", result.main_ci_status),
-                "main_fix_attempted": result.main_fix_attempted,
-            });
-
-            Ok(to_json_text(&output))
-        })
-    }
-}
 
 // ── parateresis (παρατήρησις — observation) ────────────────────────────────
 
@@ -780,7 +687,7 @@ pub fn register(
     )?;
     registry.register(qa::dokimasia_def(), Box::new(qa::DokimasiaExecutor))?;
     registry.register(qa::diorthosis_def(), Box::new(qa::DiorthosisExecutor))?;
-    registry.register(epitropos_def(), Box::new(EpitroposExecutor))?;
+    registry.register(steward::epitropos_def(), Box::new(steward::EpitroposExecutor))?;
     registry.register(
         parateresis_def(),
         Box::new(ParateresisExecutor {
@@ -819,7 +726,7 @@ mod tests {
             dispatch::dromeus_def(),
             qa::dokimasia_def(),
             qa::diorthosis_def(),
-            epitropos_def(),
+            steward::epitropos_def(),
             parateresis_def(),
         ] {
             assert_eq!(
@@ -841,7 +748,7 @@ mod tests {
             dispatch::dromeus_def(),
             qa::dokimasia_def(),
             qa::diorthosis_def(),
-            epitropos_def(),
+            steward::epitropos_def(),
             parateresis_def(),
             mathesis_def(),
             prographe_def(),
